separated front end components and added socketio
