+++ front end/package.json
@@ -17,6 +17,7 @@
     "svelte": "^3.0.0"
   },
   "dependencies": {
-    "sirv-cli": "^1.0.0"
+    "sirv-cli": "^1.0.0",
+    "socket.io-client": "^4.5.1"
   }
 }
